moved fns out of cli entry point into data prep
