+++ text_to_conll_cli.py
@@ -72,76 +72,6 @@ def print_to_conll(sentence_tuples, annotations=None, sentences=None):
             [print(annotation) for annotation in annotations[i]]
         print("\n".join(["\t".join([str(i) for i in tup]) for tup in sentence_tuples[i]])+"\n")
 
-def get_feats_from_text_tuples(text_tuples: List[List[tuple]]) -> List[List[str]]:
-    """Extract the FEATS columns from the unparsed data.
-    FEATS will exist only for text and pre-processed text inputs.
-
-    Args:
-        text_tuples (List[List[tuple]]): unparsed data
-
-    Returns:
-        List[List[str]]: the FEATS column (or _ if it does not exist)
-    """
-    try:
-        return [[col_items[5] for col_items in tup_list] for tup_list in text_tuples]
-    except:
-        import pdb; pdb.set_trace()
-
-def add_feats(text_tuples: List[List[tuple]], text_feats: List[List[str]]) -> List[List[tuple]]:
-    """Add FEATS data to the text tuples.
-    The parent list (text_tuples) is a list of sentences.
-    Each sentence is a list of tuples.
-    Each tuple represents a token.
-
-    Args:
-        text_tuples (List[List[tuple]]): list of list of tuples
-        text_feats (List[List[str]]): list of list of FEATS
-
-    Returns:
-        List[List[tuple]]: text_tuples but with the FEATS column filled
-    """
-    text_tuples_with_feats = []
-    for sentence_tuples, sentence_feats in zip(text_tuples, text_feats):
-        
-        # get first 5 and last 4 items from parsed tuple using lists, and add features.
-        # Convert the list of fields to a tuple
-        merged_tuples = [
-            tuple(list(token_tuple[:5]) + [token_feats] + list(token_tuple[6:]))
-            for token_tuple, token_feats in zip(sentence_tuples, sentence_feats)
-        ]
-        text_tuples_with_feats.append(merged_tuples)
-    return text_tuples_with_feats
-
-def string_to_tuple_list(string_of_tuples: str) -> List[tuple[str, str]]:
-    """Take a string of space-separated tuples and convert it to a tuple list.
-    Example input: '(جامعة, NOM) (نيويورك, PROP)'
-    Example output: [(جامعة, NOM), (نيويورك, PROP)]
-
-    Args:
-        string_of_tuples (str): string of tuples
-
-    Returns:
-        List(tuple[str, str]): list of token-pos tuple pairs
-    """
-    sentence_tuples = []
-    
-    # split on space, and using positive lookbehind and lookahead
-    # to detect parentheses around the space
-    for tup in re.split(r'(?<=\)) (?=\()', string_of_tuples.strip()):
-        # tup = (جامعة, NOM)
-        tup_items = tup[1:-1] # removes parens
-        form = (','.join(tup_items.split(',')[:-1])).strip() # account for comma tokens
-        pos = (tup_items.split(',')[-1]).strip()
-        sentence_tuples.append((form, pos))
-    return sentence_tuples
-
-def get_tree_tokens(tok_pos_tuples):
-    sentences = []
-    for sentence_tuples in tok_pos_tuples:
-        sentence = ' '.join([tok_pos_tuple[0] for tok_pos_tuple in sentence_tuples])
-        sentences.append(sentence)
-    return sentences
-
 def get_file_type(file_type):
     if file_type in ['conll', 'raw', 'tokenized', 'tok_tagged', 'parse_tok']:
         return file_type
